+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 Changelog
 =========
 
-Version 13.14
+Version 15.1
 =============
 
 * Refactored :meth:`IQMBackend.create_run_request` to improve user experience when using IQM specific run options.
@@ -12,13 +12,45 @@ Version 13.14
 * Introduction of `IQMBackendBase.physical_target` and `IQMBackendBase.fake_target` to represent the physical quantum architectures and a Qiskit-compatible version, respectively.        
 * Added support for ``qiskit == 1.2`` and ``qiskit-aer == 1.5``.
 
+Version 15.0
+============
+
+* Add support for classically controlled R, RX, RY, X, and Y gates.
+  `#123 <https://github.com/iqm-finland/qiskit-on-iqm/pull/123>`_
+* Remove the deprecated native operation names ``phased_rx`` and ``measurement``.
+  `#123 <https://github.com/iqm-finland/qiskit-on-iqm/pull/123>`_
+* Add support for the Reset instruction.
+  `#123 <https://github.com/iqm-finland/qiskit-on-iqm/pull/123>`_
+
+Version 14.0
+============
+
+* Use dynamic quantum architecture as transpilation target for :class:`IQMBackend`. `#124 <https://github.com/iqm-finland/iqm-client/pull/124>`_
+* Require ``iqm-client >= 20.0``. `#124 <https://github.com/iqm-finland/iqm-client/pull/124>`_
+* Disable attestations on ``gh-action-pypi-publish`` to fix failing PyPI publishing. `#124 <https://github.com/iqm-finland/iqm-client/pull/124>`_
+
+Version 13.16
+=============
+
+* Remove unnecessary build files when publishing documentation. `#122 <https://github.com/iqm-finland/iqm-client/pull/122>`_
+
+Version 13.15
+=============
+
+* User guide updated. `#120 <https://github.com/iqm-finland/qiskit-on-iqm/pull/120>`_
+
+Version 13.14
+=============
+
+* User guide and API documentation updated. `#117 <https://github.com/iqm-finland/qiskit-on-iqm/pull/117>`_
+
 Version 13.13
 =============
 
 * Adjustments needed to support Qiskit V1 that are backwards compatible with ``qiskit < 1.0``. `#114 <https://github.com/iqm-finland/qiskit-on-iqm/pull/114>`_
-* Updated Qiskit dependencies and testing to support ``qiskit >= 0.45.3 < 1.2`` and ``qiskit-aer >= 0.13 < 0.15``. 
+* Updated Qiskit dependencies and testing to support ``qiskit >= 0.45.3 < 1.2`` and ``qiskit-aer >= 0.13 < 0.15``.
 * Adjusted documentation to recommend the use of :meth:`qiskit.transpile()` or :meth:`transpile_to_IQM()` in combination with :meth:`backend.run()` instead of using :meth:`execute()`.
-* Added a DeprecationWarning for use of ``qiskit < 1.0``. Users are encouraged to read the `Qiskit migration guide <https://docs.quantum.ibm.com/migration-guides>`_ to see how they need to change their code. 
+* Added a DeprecationWarning for use of ``qiskit < 1.0``. Users are encouraged to read the `Qiskit migration guide <https://docs.quantum.ibm.com/migration-guides>`_ to see how they need to change their code.
 
 Version 13.12
 =============